backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.32.3
openai==1.54.3
python-dotenv==1.1.1
pydantic==2.9.2
pydantic[email]==2.9.2